Introduce more flexible State field
Since a state is effecively just a string/flag defined by the kernel, we
cannot always rely on only getting what we have defined on our side.
Therefore, getting a value that we do not know should not be a hard
error.
Instead, return it as Undefined, along with the information that the OS
returned to us.

BREAKING: See BREAKING.md file for details.

ALSO: Reworked OpenBSD code so it is not *so* awkward.